[BE #17] Refactor: 패키지 오타 수정
패키지명 오타 수정하였음
